moved all slack related functions into slackchannel package, removed initialResources struct
